desert 게시판 Pagination 완료 및 비동기 댓글 리스트 형식 수정

+++ src/main/java/com/example/jmt/desert/service/CommentDesertService.java
@@ -45,7 +45,15 @@ public class CommentDesertService {
     }
 
     // 삭제
-    public void deleteComment(Long commentId) {
+    public void deleteComment(Long commentId, User user) {
+
+        CommentDesert commentDesert = commentDesertRepository.findById(commentId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 댓글입니다."));
+
+        if (!commentDesert.getUser().equals(user)) {
+            throw new IllegalArgumentException("삭제 권한이 없습니다.");
+        }
+
         commentDesertRepository.deleteById(commentId);
     }
 }
